restructure to make library more tree-shakable without importing built code
fix lint
